Renamed function module to get-required-information-warnings
